4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -48,11 +48,21 @@ blogsRouter.post('/', async (request, response, next) => {
   })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
-      await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
-    } catch(error) {
-      next(error)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid '})
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if ( user._id.toString() === blog.user.toString() ) {
+      try {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } catch(error) {
+        next(error)
+      }
+    } else {
+      return response.status(401).json({ error: 'wrong user' })
     }
 })
